requirements.txt
langchain
langchain_openai
openai
langchain_community
bs4
pypdf
chromadb
faiss-cpu
sentence_transformers
crewai
crewai_tools
langchain_google_genai
load_dotenv
Flask
reportlab
google-generativeai
fpdf
PyPDF2
phidata
ipywidgets
langchain-ollama
yfinance
pandas
numpy
plotly
tensorflow
scikit-learn
datetime
newsapi-python
